backend/requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Authentication
django-allauth==0.57.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment variables
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
pytz==2023.3
